revisi

+++ src/components/sidebar/Sidebar.tsx
@@ -5,6 +5,18 @@ import { LogOut } from 'lucide-react'
 import { useState } from 'react'
 import { ResizablePanel } from '../ui/resizable'
 import SidebarItem from './SidebarItem'
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger
+} from '@/components/ui/alert-dialog'
+import { buttonVariants } from '../ui/button'
 
 const Sidebar = () => {
   const [isCollapsed, setIsCollapsed] = useState(false)
@@ -41,7 +53,7 @@ const Sidebar = () => {
         setIsCollapsed(false)
         document.cookie = `react-resizable-panels:collapsed=${JSON.stringify(false)}`
       }}
-      className={cn('flex flex-col', {
+      className={cn('flex flex-col bg-primary-blue-700 text-white sidebar', {
         'min-w-[3.125rem] transition-all duration-300 ease-in-out items-center': isCollapsed
       })}
     >
@@ -64,17 +76,35 @@ const Sidebar = () => {
           ))}
         </nav>
         <div className='px-2 lg:px-4'>
-          <div
-            role='button'
-            className='flex items-center gap-3 rounded-lg px-3 pb-10 transition-all [&>svg]:size-4 group-data-[mobile=true]:[&>svg]:size-5 text-sm text-red-500'
-            onClick={() => {
-              localStorage.clear()
-              location.reload()
-            }}
-          >
-            <LogOut />
-            {!isCollapsed && 'Logout'}
-          </div>
+          <AlertDialog>
+            <AlertDialogTrigger asChild>
+              <div
+                role='button'
+                className='flex items-center gap-3 rounded-lg px-3 pb-10 transition-all [&>svg]:size-4 group-data-[mobile=true]:[&>svg]:size-5 text-sm text-red-500'
+              >
+                <LogOut />
+                {!isCollapsed && 'Logout'}
+              </div>
+            </AlertDialogTrigger>
+            <AlertDialogContent>
+              <AlertDialogHeader>
+                <AlertDialogTitle>Keluar Aplikasi</AlertDialogTitle>
+                <AlertDialogDescription>Anda yakin ingin keluar?</AlertDialogDescription>
+              </AlertDialogHeader>
+              <AlertDialogFooter>
+                <AlertDialogCancel>Batal</AlertDialogCancel>
+                <AlertDialogAction
+                  className={buttonVariants({ variant: 'destructive' })}
+                  onClick={() => {
+                    localStorage.clear()
+                    location.reload()
+                  }}
+                >
+                  Keluar
+                </AlertDialogAction>
+              </AlertDialogFooter>
+            </AlertDialogContent>
+          </AlertDialog>
         </div>
       </div>
     </ResizablePanel>
